+++ core-ng/src/main/java/core/framework/module/SchedulerConfig.java
@@ -41,7 +41,7 @@ public final class SchedulerConfig extends Config {
     }
 
     public void timeZone(ZoneId zoneId) {
-        if (triggerAdded) throw new Error("schedule timeZone must be configured before adding trigger");
+        if (triggerAdded) throw new Error("schedule().timeZone() must be configured before adding trigger");
         if (zoneId == null) throw new Error("zoneId must not be null");
         scheduler.clock = Clock.system(zoneId);
     }