+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 ## Change log
-### 7.3.1 (12/20/2019 - )
+### 7.3.1 (12/20/2019 - 1/15/2020)
 * es: update to 7.5.1
 * kafka: use null as key if not specified, since from kafka 2.4.0 it supports sticky partitioning, refer to refer to org.apache.kafka.clients.producer.internals.DefaultPartitioner
-         use 950k as max message size for producer
-* log-collector: validate event.info size with max=800k         
+         set producer max request size to match broker default setting 
+* log-collector: validate event.info size with max=900k         
 
 ### 7.3.0 (12/11/2019 - 12/18/2019)
 * kafka: update to 2.4.0